Allow to override the emacs package used for running the script
Takes the emacs package as an argument to the outermost closure, so
that we can use the same emacs package to run the script as we
install.

We also change to take everything we use as an explicit argument (no
longer taking pkgs), since this module is designed to be loaded with
callPackage anyway.